ai-engine/requirements.txt
# Core FastAPI Dependencies
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database
motor>=3.3.2,<4.0.0
pymongo>=4.6.1,<5.0.0

#  LangChain & Related
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-openai==0.2.0
langchain-google-genai==2.0.0
langgraph==0.2.40



# LlamaIndex
llama-index>=0.9.10,<1.0.0
llama-hub>=0.0.41,<0.1.0

# Vector Stores
chromadb>=0.4.18,<0.5.0

# Graph Databases
neo4j>=5.14.0,<6.0.0
networkx>=3.2.1,<4.0.0

# LLM Providers
openai>=1.6.1,<2.0.0
google-generativeai>=0.7.0,<0.8.0

# Hugging Face Integration
transformers>=4.36.0,<5.0.0
torch>=2.1.1,<2.2.0
torchvision>=0.16.1,<0.17.0
accelerate>=0.25.0,<1.0.0
bitsandbytes>=0.41.1,<1.0.0
einops>=0.7.0,<1.0.0
sentencepiece>=0.1.99,<1.0.0

# Document Processing
unstructured>=0.10.30,<0.11.0
python-magic>=0.4.27,<0.5.0
pdf2image>=1.16.3,<2.0.0
pytesseract>=0.3.10,<1.0.0
python-docx>=0.8.11,<1.0.0
python-pptx>=0.6.22,<1.0.0
openpyxl>=3.1.2,<4.0.0

# Utilities
tenacity>=8.2.3,<9.0.0
aiohttp>=3.9.1,<4.0.0
requests>=2.31.0,<3.0.0
numpy>=1.26.0,<2.0.0
pandas>=2.1.4,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0

# Development & Testing
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.23.2,<1.0.0
httpx>=0.25.2,<1.0.0
black>=23.11.0,<24.0.0
isort>=5.13.2,<6.0.0
mypy>=1.7.1,<2.0.0
pytest-cov>=4.1.0,<5.0.0
 
beautifulsoup4